--- src/test/java/br/edu/ufcg/computacao/eureca/backend/util/RequestFactory.java
@@ -1,32 +1,0 @@
-package br.edu.ufcg.computacao.eureca.backend.util;
-
-import br.edu.ufcg.computacao.eureca.backend.api.http.CommonKeys;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.RequestBuilder;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-
-public class RequestFactory {
-
-    public static RequestBuilder createRequestBuilder(HttpMethod method, String url, HttpHeaders headers, String content) {
-        switch (method) {
-            case GET:
-                return MockMvcRequestBuilders.get(url)
-                        .headers(headers)
-                        .accept(MediaType.APPLICATION_JSON)
-                        .content(content)
-                        .contentType(MediaType.APPLICATION_JSON);
-            case POST:
-                return null;
-            default:
-                return null;
-        }
-    }
-
-    public static HttpHeaders getTokenHeaders() {
-        HttpHeaders headers = new HttpHeaders();
-        headers.set(CommonKeys.AUTHENTICATION_TOKEN_KEY, "fake-token");
-        return headers;
-    }
-}
